fix source file names
