+++ Makefile
@@ -1,0 +1,4 @@
+TARGET	= metro
+
+INCLUDE	= include
+SOURCES	=
