+++ src/MetroDriver/Evaluator/evaluator.cc
@@ -3,6 +3,28 @@
 #include "GC.h"
 
 namespace Metro {
+  Object** Evaluator::eval_lvalue(AST::Base* ast) {
+    using ASTKind = AST::Kind;
+
+    switch( ast->kind ) {
+      case ASTKind::Variable: {
+        auto x = (AST::Variable*)ast;
+
+        switch( x->defined->kind ) {
+          case ASTKind::Argument:
+            return &((AST::Argument*)x->defined)->value;
+
+          case ASTKind::VarDefine:
+            return &((AST::VarDefine*)x->defined)->value;
+        }
+
+        break;
+      }
+    }
+
+    crash;
+  }
+
   Object* Evaluator::eval(AST::Base* ast) {
     using AST::Kind;
